Got Checkout working!
Checkout is currently working as intended! It checks if a user has met/exceeded allowed books, checks if the book is available and not reserved by another person, and if all checks pass, an entry is put into the "checkouts" table and the available copies in "books" is subtracted by 1. 

I will get Returns working before the end of today.