Updated Checkouts and Returns!
Checkouts and Returns now interacts with the notifications table. Upon checkout, if the user has the checked-out book on their notifications, it will remove that row from the notifications table. Upon return, the return function checks if the book is reserved or if people are looking for notifications. If it's reserved, the reserve will be emailed using the "mail()" function. If not, than everyone who wants to be notified will be emailed. 

However, you'll have to set up mailEnable and configure php.ini to get this to work.